requirements.txt
playwright
python-dotenv
pandas
sqlite-utils
requests
beautifulsoup4
